+++ CHANGELOG.md
@@ -3,12 +3,20 @@ All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [1.1.0] - 2020-08-29
+## Added
+- Restore scene after exit play mode.
+- Support for Unity 2020.1.
+
+## Changed
+- Shortcut name is now displayed as tooltip.
+
 ## [1.0.1] - 2019-06-12
 This is the first release of *Scene Inspector* package via Unity Package Manager.
 
 ### Added
-- Create new scene option
-- Pin scene shortcut to toolbar
+- Create new scene option.
+- Pin scene shortcut to toolbar.
 
 ### Changed
 - UX tweaks